PostController, MySql, User
- Ahora se comprueba el número de filas afectadas al realizar un DELETE.
- Se agrego un lista con los ultimos 5 posts y 5 comentarios a la pagina
de inicio del panel de administración.
- Se actualizo del controlador de POST para implementar los modulos de
post-categorías y post-etiquetas.
- Se actualizo las vistas de POST.
- Se actualizo los modulos DELETE, se agregaron funciones para preparar
la consulta.
- Agregado funciones estaticas en el modelo de comentarios.
- Ahora al mostrar la lista de usuarios, en el panel de administración,
los botones de acciones ya no se muestran pero aun funcionan si se
escribe la URL.
- Otros cambios...

+++ app/views/admin/user.php
@@ -34,9 +34,11 @@
                         foreach ($data['users'] as $user) {
                             $output = '<tr>';
                             $output .= '<td class="options">';
-                            $output .= '<a class="btnAction-sm btn btn-primary" href="' . $data['siteUrl'] . 'admin/user/update/' . $user->getID() . '" title="Editar"><span class="glyphicon glyphicon-edit"></span></a> ';
-                            $output .= '<a class="btnAction-sm btn btn-danger" href="' . $data['siteUrl'] . 'admin/user/delete/' . $user->getID() . '" title="Editar"><span class="glyphicon glyphicon-remove-sign"></span></a> ';
+                            if ($_SESSION['usernameID'] != $user->getID()) {
+                                $output .= '<a class="btnAction-sm btn btn-primary" href="' . $data['siteUrl'] . 'admin/user/update/' . $user->getID() . '" title="Editar"><span class="glyphicon glyphicon-edit"></span></a> ';
+                                $output .= '<a class="btnAction-sm btn btn-danger" href="' . $data['siteUrl'] . 'admin/user/delete/' . $user->getID() . '" title="Editar"><span class="glyphicon glyphicon-remove-sign"></span></a> ';
 //                            $output .= '<button class="btnAction btnAction-sm btn btn-danger" data-action="" title="Borrar"><span class="glyphicon glyphicon-remove-sign"></span></button></td>';
+                            }
                             $output .= '</td>';
                             $output .= '<td>' . $user->getUserLogin() . '</td>';
                             $output .= '<td>' . $user->getUserName() . '</td>';
